[마케터 분석] 분석페이지 -> 대시보드 대체로인한 구조변경
