+++ client/src/organisms/marketer/Dashboard/banner/BannerDescForm.jsx
@@ -4,7 +4,7 @@ import { withStyles } from '@material-ui/core/styles';
 import {
   Grid, FormControl, InputLabel, Input, FormHelperText
 } from '@material-ui/core';
-import CustomButton from '../../../../../atoms/CustomButtons/Button';
+import CustomButton from '../../../../atoms/CustomButtons/Button';
 
 const formStyle = theme => ({
   imgPreview: {
