+++ client/src/organisms/marketer/Dashboard/campaign/CampaignDeleteConfirmDialog.jsx
