--- client/src/organisms/marketer/Dashboard/sub/GraphConsole.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-export default function GraphConsole() {
-  return (
-    <div>
-      <div>
-        그래프 제어 관련 기능은
-      </div>
-      <div>
-      차후 제공될 예정입니다.
-      </div>
-    </div>
-  );
-}
