+++ client/src/organisms/marketer/Dashboard/table/TableSubComponent.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function TableSub() {
+  return <div />;
+}
