+++ client/src/pages/marketer/Dashboard.jsx
@@ -1,73 +1,120 @@
 import React from 'react';
-// @material-ui/core
-import Hidden from '@material-ui/core/Hidden';
-// core ../../../atoms
-import GridContainer from '../../atoms/Grid/GridContainer';
-import GridItem from '../../atoms/Grid/GridItem';
-import ValueChart from '../../organisms/marketer/Dashboard/ValueChart';
-import CashCard from '../../organisms/marketer/Dashboard/CashCard';
-import StatusBar from '../../organisms/marketer/Dashboard/CampaignOnOffSwitch';
-import CampaignCreate from '../../organisms/marketer/Dashboard/CampaignCreate';
+import makeStyles from '@material-ui/core/styles/makeStyles';
+import Grid from '@material-ui/core/Grid';
+import Grow from '@material-ui/core/Grow';
+
+import BannerList from '../../organisms/marketer/Dashboard/BannerList';
 import CampaignList from '../../organisms/marketer/Dashboard/CampaignList';
-
+import CanvasForChart from '../../organisms/marketer/Dashboard/CanvasForChart';
+import DescCard from '../../organisms/marketer/Dashboard/DescCard';
+import OnOffSwitch from '../../organisms/marketer/Dashboard/OnOffSwitch';
+import ReportLoading from '../../organisms/marketer/Dashboard/ReportLoading';
+import IssueTable from '../../organisms/marketer/Dashboard/IssueTable';
+// hooks
 import useFetchData from '../../utils/lib/hooks/useFetchData';
-import useToggle from '../../utils/lib/hooks/useToggle';
 
-const Dashboard = () => {
-  const campaignCreateMode = useToggle();
-  const marketerOnOffData = useFetchData('/api/dashboard/marketer/onoff');
-  const marketerProfileData = useFetchData('/api/dashboard/marketer/profile');
-  const campaignData = useFetchData('/api/dashboard/marketer/campaign/list');
-  const bannerData = useFetchData('/api/dashboard/marketer/banner');
+const useStyles = makeStyles(theme => ({
+  root: {
+    [theme.breakpoints.up('xl')]: {
+      margin: '0px 160px'
+    }
+  }
+}));
+
+export default function Dashboard() {
+  const classes = useStyles();
+  const campaignData = useFetchData('/api/dashboard/marketer/report/new');
+  const onOffData = useFetchData('/api/dashboard/marketer/onoff');
+  const normalData = useFetchData('/api/dashboard/marketer/report/normal');
+  const creatorsData = useFetchData('/api/dashboard/marketer/report/creators');
+  const bannerData = useFetchData('/api/dashboard/marketer/banner/all');
+  const valueChartData = useFetchData('/api/dashboard/marketer/campaign/chart');
+  const broadCreatorData = useFetchData('/api/dashboard/marketer/report/broadcast/creator');
 
   return (
-    <div>
-      {campaignCreateMode.toggle && (
-        <GridContainer>
-          <GridItem xs={12} xl={12}>
-            <CampaignCreate handleCampaignCreateMode={campaignCreateMode.handleToggle} />
-          </GridItem>
-        </GridContainer>
-      )}
+    <div className={classes.root}>
+      {(normalData.loading || campaignData.loading
+        || onOffData.loading || creatorsData.loading
+        || bannerData.loading || valueChartData.loading) ? (
+          <ReportLoading />
+        ) : (
+          <Grid container spacing={2}>
+            {normalData.payload && campaignData.payload
+            && creatorsData.payload && bannerData.payload
+            && valueChartData.payload && (
+            <Grid item xs={12}>
+              <Grid container spacing={2}>
+                <Grid item xs={12} sm={6} lg={3}>
+                  <Grow in timeout={{ enter: 300 }}>
+                    <DescCard data={{
+                      title: '광고 캐시 잔액', value: normalData.payload.cashAmount, unit: '원'
+                    }}
+                    />
+                  </Grow>
+                </Grid>
+                <Grid item xs={12} sm={6} lg={3}>
+                  <Grow in timeout={{ enter: 700 }}>
+                    <DescCard data={{
+                      title: '총 소진 비용', value: normalData.payload.spendAll, unit: '원'
+                    }}
+                    />
+                  </Grow>
+                </Grid>
+                <Grid item xs={12} sm={6} lg={3}>
+                  <Grow in timeout={{ enter: 1100 }}>
+                    <DescCard data={{
+                      title: '운용중 캠페인',
+                      value: campaignData.payload.filter(c => c.onOff === 1).length,
+                      unit: '캠페인'
+                    }}
+                    />
+                  </Grow>
+                </Grid>
+                <Grid item xs={12} sm={6} lg={3}>
+                  <Grow in timeout={{ enter: 1500 }}>
+                    <DescCard data={{
+                      title: '송출크리에이터수', value: creatorsData.payload.length, unit: '명'
+                    }}
+                    />
+                  </Grow>
+                </Grid>
+              </Grid>
+            </Grid>
+            )}
 
-      {!campaignCreateMode.toggle && (
-        <div>
-          <GridContainer>
-            <GridItem xs={12} md={12} xl={12}>
-              <StatusBar onOffData={marketerOnOffData} />
-            </GridItem>
-          </GridContainer>
+            <Grid item xs={12} md={6} lg={3}>
+              <Grid container spacing={2}>
+                <Grid item xs={12}>
+                  <OnOffSwitch onOffData={onOffData} />
+                </Grid>
+                <Grid item xs={12}>
+                  <CampaignList campaignData={campaignData} />
+                </Grid>
+                <Grid item xs={12}>
+                  <BannerList bannerData={bannerData} />
+                </Grid>
+              </Grid>
+            </Grid>
 
-          <GridContainer>
-            <GridItem xs={12} xl={3}>
-              <CashCard
-                marketerProfileData={marketerProfileData}
-                campaignData={campaignData}
-                bannerData={bannerData}
-              />
-            </GridItem>
-            <Hidden xsDown>
-              <GridItem sm={12} xl={9}>
-                <ValueChart />
-              </GridItem>
-            </Hidden>
-          </GridContainer>
+            <Grid item xs={12} md={6} lg={9}>
+              <Grid container spacing={2}>
+                <Grid item xs={12}>
+                  <CanvasForChart
+                    valueChartData={valueChartData}
+                    creatorsData={creatorsData}
+                    broadCreatorData={broadCreatorData}
+                  />
+                </Grid>
+                <Grid item xs={12}>
+                  <IssueTable />
+                </Grid>
 
-          <GridContainer>
-            <GridItem xs={12} xl={12}>
-              <CampaignList
-                campaignData={campaignData}
-                handleCampaignCreateMode={campaignCreateMode.handleToggle}
-                bannerData={bannerData}
-              />
-            </GridItem>
-          </GridContainer>
+              </Grid>
+            </Grid>
 
-        </div>
-      )}
 
+          </Grid>
+        )}
     </div>
   );
-};
-
-export default Dashboard;
+}
